Improved: code to handle parties adding removing action call at modal level (#12)
